requirements.txt
Flask
gunicorn
azure-storage-blob
azure-mgmt-storage
flask-wtf
werkzeug
requests
waitress